refactor: refactor response processing
